backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.32.1

# Database
sqlalchemy[asyncio]==2.0.36
alembic==1.14.0
asyncpg==0.30.0
psycopg2-binary==2.9.10
pgvector==0.3.6

# Config
python-dotenv==1.0.1
pydantic-settings==2.7.0

# File uploads
python-multipart==0.0.20

# Cloud storage
boto3==1.35.86

# AI / LLM
anthropic==0.40.0
openai==1.57.4

# Auth
unkey==0.0.2

# Async workers
modal==0.73.35

# Transcription
assemblyai==0.36.0

# NLP
spacy==3.8.7

# Testing
pytest==8.3.4
httpx==0.28.1
stripe==11.4.0
PyJWT==2.10.1
cryptography==43.0.3
